+++ gradle-plugin-testing-junit/src/testFixtures/java/com/palantir/example/MultipleGradleVersionsFixtureTest.java
@@ -1,0 +1,39 @@
+/*
+ * (c) Copyright 2025 Palantir Technologies Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.palantir.example;
+
+import com.palantir.gradle.testing.execution.Gradlew;
+import com.palantir.gradle.testing.junit.GradlePluginTests;
+import com.palantir.gradle.testing.project.RootProject;
+import org.junit.jupiter.api.Test;
+
+@GradlePluginTests
+public class MultipleGradleVersionsFixtureTest {
+    @Test
+    @SuppressWarnings("checkstyle:RegexpSinglelineJava")
+    void test(Gradlew gradlew, RootProject rootProject) {
+        rootProject
+                .buildFile()
+                .append(
+                        """
+        import org.gradle.util.GradleVersion
+        println "GradleVersion: ${GradleVersion.current().version}"
+        """);
+
+        throw new RuntimeException(gradlew.withArgs().buildSuccessfully().output());
+    }
+}
